server/python_flask/requirements.txt
flask
flask_restful
flask_sqlalchemy
tqdm
requests
pandas
mysqlclient
geopy
flask-cors
scipy
